backend/requirements.txt
asgiref==3.6.0
decouple==0.0.7
distlib==0.3.6
Django==4.1.3
djangorestframework==3.14.0
filelock==3.10.7
platformdirs==3.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.3
tzdata==2023.3
virtualenv==20.21.0
django-cors-headers==3.13.0
reportlab==3.6.12
